Vdb 1788 audit execute errors (#178)
* Remove redundant error logging

These errors are returned from StorageWatched.Execute and being logged
at the top level execute command. Removing this additional logging will
make for cleaner log messages.

in the top level command.

* Change log extract errors to Warn from Error

This particular log can be changed to a Warn because we have a withRetry
method in the EventWatcher will will retry a set number of times, and if
it continues to fail an error will be logged on line 95 of the event
watcher.

* Downgrade log errors to warns in EventWatcher, Extractor and Delegator

Since these errors are being returned, they'll be logged at the top
level execute command, and probably don't need to be logged as errors
here as well.

* Downgrade log errors to warns in event Transformer

Since these errors are being returned, they'll be logged at the top
level execute command, and probably don't need to be logged as errors
here as well.